refactor: move db models from shared to backend DB
