+++ packages/backend/src/db/models/project/DBFileMetadata.ts
@@ -1,6 +1,6 @@
-import { DBDatedData } from "./DBDatedData";
-import { VersionRelation } from "@shared/dbModels/project/DBVersion";
-import { DBFileData } from "@shared/dbModels/project/DBFileData";
+import { DBDatedData } from "@db/models/project/DBDatedData";
+import { VersionRelation } from "@db/models/project/DBVersion";
+import { DBFileData } from "@db/models/project/DBFileData";
 
 // table name: files
 export interface DBFileMetadata extends VersionRelation, DBDatedData {
