+++ packages/backend/src/db/sqlHelpers/dbDates.ts
@@ -1,7 +1,4 @@
-import {
-  DBDatedData,
-  DBSoftDeletable,
-} from "@shared/dbModels/project/DBDatedData";
+import { DBDatedData, DBSoftDeletable } from "@db/models/project/DBDatedData";
 import { DatedData } from "@shared/domain/readModels/project/DatedData";
 import moment from "moment";
